fix: accept utcOffset as JSON string or number in quick_add
iOS Shortcuts sends all JSON dictionary values as strings. Added
FlexibleInt16 type that accepts both "60" and 60 for the utcOffset
field.

+++ pkg/api/transactions.go
@@ -1200,7 +1200,7 @@ func (a *TransactionsApi) TransactionQuickAddHandler(c *core.WebContext) (any, *
 		Type:            models.TRANSACTION_TYPE_EXPENSE,
 		CategoryId:      categoryId,
 		Time:            quickAddReq.Time,
-		UtcOffset:       quickAddReq.UtcOffset,
+		UtcOffset:       int16(quickAddReq.UtcOffset),
 		SourceAccountId: accountId,
 		SourceAmount:    amountCents,
 		Comment:         quickAddReq.Merchant,
